renaming content to datastore
