+++ nexus-repository-terraform/src/main/java/org/sonatype/nexus/plugins/terraform/datastore/TerraformContentFacet.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.plugins.terraform.content;
+package org.sonatype.nexus.plugins.terraform.datastore;
 
 import java.io.IOException;
 import java.util.Optional;
