+++ nexus-repository-terraform/src/main/java/org/sonatype/nexus/plugins/terraform/datastore/internal/recipe/TerraformContentHandler.java
@@ -10,14 +10,14 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.plugins.terraform.content.internal.recipe;
+package org.sonatype.nexus.plugins.terraform.datastore.internal.recipe;
 
 import javax.annotation.Nonnull;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
 import org.sonatype.goodies.common.ComponentSupport;
-import org.sonatype.nexus.plugins.terraform.content.TerraformContentFacet;
+import org.sonatype.nexus.plugins.terraform.datastore.TerraformContentFacet;
 import org.sonatype.nexus.plugins.terraform.internal.AssetKind;
 import org.sonatype.nexus.plugins.terraform.internal.Attributes.TerraformAttributes;
 import org.sonatype.nexus.plugins.terraform.internal.util.TerraformPathUtils;
@@ -92,6 +92,7 @@ public class TerraformContentHandler
   @Nonnull
   private String contentPath(final Context context) {
     TokenMatcher.State state = context.getAttributes().require(TokenMatcher.State.class);
+    System.out.println("DABRAAAAA: " + state.getTokens());
     String path = state.getTokens().get("path");
     checkState(path != null, "Missing token: path");
