+++ nexus-repository-terraform/src/main/java/org/sonatype/nexus/plugins/terraform/datastore/internal/store/TerraformAssetBlobDAO.java
@@ -10,15 +10,15 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.plugins.terraform.content.internal.store;
+package org.sonatype.nexus.plugins.terraform.datastore.internal.store;
 
-import org.sonatype.nexus.repository.content.store.ComponentDAO;
+import org.sonatype.nexus.repository.content.store.AssetBlobDAO;
 
 /**
  * @since 0.0.6
  */
-public interface TerraformComponentDAO
-        extends ComponentDAO
+public interface TerraformAssetBlobDAO
+        extends AssetBlobDAO
 {
   // nothing to add...
 }
