+++ nexus-repository-terraform/src/main/java/org/sonatype/nexus/plugins/terraform/datastore/internal/store/TerraformContentRepositoryDAO.java
@@ -10,21 +10,15 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.plugins.terraform.content.internal.browse;
+package org.sonatype.nexus.plugins.terraform.datastore.internal.store;
 
-import javax.inject.Named;
-
-import org.sonatype.nexus.plugins.terraform.internal.TerraformFormat;
-import org.sonatype.nexus.repository.content.browse.store.FormatBrowseModule;
+import org.sonatype.nexus.repository.content.store.ContentRepositoryDAO;
 
 /**
- * Configures the browse bindings for the 'terraform' format.
- *
  * @since 0.0.6
  */
-@Named(TerraformFormat.NAME)
-public class TerraformBrowseModule
-    extends FormatBrowseModule<TerraformBrowseNodeDAO>
+public interface TerraformContentRepositoryDAO
+        extends ContentRepositoryDAO
 {
   // nothing to add...
 }
