+++ nexus-repository-terraform/src/main/java/org/sonatype/nexus/plugins/terraform/datastore/internal/store/TerraformStoreModule.java
@@ -1,6 +1,6 @@
 /*
  * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2008-present Sonatype, Inc.
+ * Copyright (c) 2022-present Sonatype, Inc.
  * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
  *
  * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
@@ -10,9 +10,25 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-@FeatureFlag(name = DATASTORE_ENABLED)
-package org.sonatype.nexus.plugins.terraform.content;
+package org.sonatype.nexus.plugins.terraform.datastore.internal.store;
 
-import org.sonatype.nexus.common.app.FeatureFlag;
+import javax.inject.Named;
 
-import static org.sonatype.nexus.common.app.FeatureFlags.DATASTORE_ENABLED;
+import org.sonatype.nexus.repository.content.store.FormatStoreModule;
+import org.sonatype.nexus.plugins.terraform.internal.TerraformFormat;
+
+/**
+ * Configures the content store bindings for the 'terraform' format.
+ *
+ * @since 0.0.6
+ */
+@Named(TerraformFormat.NAME)
+public class TerraformStoreModule
+    extends FormatStoreModule<TerraformContentRepositoryDAO,
+        TerraformComponentDAO,
+        TerraformAssetDAO,
+        TerraformAssetBlobDAO,
+        TerraformSearchDAO>
+{
+  // nothing to add...
+}
